Closes trac:2008: Move CID overflow detection to 1.3
Submitted by brbarrett, reviewed by rhc, no RM-approval required

Patch to cleanup the changesets for this CMR - silences compiler
warnings about comparing signed vs unsigned values.

This commit was SVN r21906.

The following Trac tickets were found above:
  Ticket 2008 --> https://svn.open-mpi.org/trac/ompi/ticket/2008

+++ ompi/mca/pml/csum/pml_csum.c
@@ -196,7 +196,7 @@ int mca_pml_csum_add_comm(ompi_communicator_t* comm)
     }
 
     /* should never happen, but it was, so check */
-    if (comm->c_contextid > mca_pml_csum.super.pml_max_contextid) {
+    if (comm->c_contextid > (uint32_t) mca_pml_csum.super.pml_max_contextid) {
         OBJ_RELEASE(pml_comm);
         return OMPI_ERR_OUT_OF_RESOURCE;
     }
